Fixed Issue with Font Allocation
-Previously a mistake was made which led to character 5 in the font
being empty and allocated into the next character slot. This caused 5 to
not appear and 6 to appear as 5.

This has been fixed.
A future change being considered is instead of individual writeMemory
calls, instead a range can be added to reduce the amount of calls and
reduce chance for a mistake.